Basic embedded support
If a struct has the 'notable' flag in its magic doc tag, no mapper or
table will be generated for it, only a few constant names.

If a struct embed another, the embedded struct fields will be added to
the former, using the embedded struct constants as much as possible.

Caveat: the embedded struct MUST be defined in the same file.